fixes for the unit and integration tests.  as expected, most of the gcc tests failed.  should i require a function to be in scope to call it?
git-svn-id: http://k-framework.googlecode.com/svn/trunk@1408 82a86efa-2583-11df-8ea0-49ba94de29ab

+++ examples/inProgress/C/tests/unitTests/testStdarg.c
@@ -70,7 +70,7 @@ int testDifferent(int x, ...){
 	int y6a = y6 * 10000;
 	int y7a = y7 * 10000;
 	struct s mys = va_arg(argp, struct s);
-	printf("%d, %d, %d, %d, %d, %d, %d, %d, %d\n", y1, y2, y3, y4, y5, y6a, y7a, (&mys)->x, (&mys)->y);
+	printf("%d, %d, %d, %d, %d, %d, %d, %d, %d\n", y1, y2, y3, (int)y4, (int)y5, y6a, y7a, (&mys)->x, (&mys)->y);
 	va_end(argp);
 	return 0;
 }
